Renamed a few files to better fit the code style used previously
