+++ src/pages/Payment/style.jsx
